backend/requirements.txt
# Core FastAPI and web framework
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy
alembic

psycopg2-binary

# AI Services
google-generativeai

# Speech Recognition
SpeechRecognition
pydub
pyaudio

# Data processing
pydantic
python-jose[cryptography]

# Development utilities
python-dotenv

reportlab
Pillow